backend/requirements.txt
djangorestframework==3.14.0
Django==4.2.7
djangorestframework-simplejwt==5.3.0
django-allauth==0.59.0
channels==4.0.0
channels-redis==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-cors-headers==4.0.0
uvicorn==0.24.0.post1
daphne==4.0.0
asgiref==3.7.2
whitenoise==6.6.0
django-jazzmin==2.6.0
Pillow==10.4.0

